pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marketing-openenv"
version = "1.0.0"
description = "OpenEnv marketing campaign optimization environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core>=0.2.0",
  "pydantic>=2.8.0",
  "numpy>=2.0.0",
  "openai>=1.40.0",
  "groq>=0.4.0",
  "fastapi>=0.112.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
